requirements.txt
Django==2.2.2
DjangoRestFramework==3.9.4